+++ bootique-jdbc-test/src/main/java/io/bootique/jdbc/test/IdentifierQuotationStrategy.java
@@ -1,0 +1,9 @@
+package io.bootique.jdbc.test;
+
+/**
+ * @since 0.14
+ */
+public interface IdentifierQuotationStrategy {
+
+    String quoted(String bareIdentifier);
+}